Made LLStore a LOT faster
I was using a process monitoring method so the Store still had time to process/update itself, but I also used it in the Loading form that needs no interaction. So I changed it to directly execute, 7z to extract .apz etc, it now takes 14 seconds to start, down from 2 minutes and  33 before the optimisations.