Moved MakeUSB to /LastOS/ instead of on the ISO as it wont run off ISO
I am moving the MakeUSB to the installed OS, which means the only way to run it will be once the OS is installed, weird I know, but I've not found a better way yet, I'll just document it and users can find alternative ways and I'll offer it as a download they can run on existing Debian Linux OS's etc. Maybe.